+++ src/_init.lua
@@ -83,5 +83,6 @@ Addon.UI = {
 
     Profiles = {}
   },
+  ItemWindow = {},
   MerchantButton = {}
 }